refactor strconv.ParseUint(c.Para refactor srp manage and workflow
